Add remove nil contract
